+++ src/main/java/tachyon/WorkerServiceHandler.java
@@ -141,14 +141,14 @@ public class WorkerServiceHandler implements WorkerService.Iface {
     addId(fileId, fileSizeBytes);
     mUsers.addOwnBytes(userId, - fileSizeBytes);
     mMasterClient.worker_cachedFile(mWorkerInfo.getId(), mWorkerInfo.getUsedBytes(), fileId,
-        (int) fileSizeBytes);
+        fileSizeBytes);
   }
 
   private void addFoundPartition(int fileId, long fileSizeBytes)
       throws FileDoesNotExistException, SuspectedFileSizeException, TException {
     addId(fileId, fileSizeBytes);
     mMasterClient.worker_cachedFile(mWorkerInfo.getId(), mWorkerInfo.getUsedBytes(), fileId,
-        (int)fileSizeBytes);
+        fileSizeBytes);
   }
 
   public void checkStatus() {
